requirements.txt
requests
psycopg2-binary
sqlalchemy
python-dotenv
jupyter
pandas
